refactor(tests): Fix ElementId usage in dependency tests
- Fixed all dependency.rs tests to use ElementId::new() instead of raw integers
- Fixed test_harness.rs hook tests to pass &mut ctx parameter
- Reduced errors from 193 to 145 (48 errors fixed)

Progress:
- ✅ Fixed 10 method argument errors in test_harness.rs
- ✅ Fixed 36 ElementId errors in dependency.rs
- 📊 Total progress: 237 → 145 errors (92 fixed, -39%)

+++ crates/flui_core/src/testing/helpers.rs
@@ -21,6 +21,6 @@ pub fn test_hook_context() -> HookContext {
 /// Create a test HookContext with a custom component ID
 pub fn test_hook_context_with_id(id: usize) -> HookContext {
     let mut ctx = HookContext::new();
-    ctx.begin_component(ComponentId(id));
+    ctx.begin_component(ComponentId(id as u64));
     ctx
 }
